requirements.txt
django==4.*
djangorestframework==3.*
google-auth==2.*
jira==3.*
